Move reused Metricbeat functionality to helper package.
