+++ metricbeat/helper/socket/user_cache.go
